break down analyzer into separate files
