+++ .papi/descriptors/package.json
@@ -1,5 +1,5 @@
 {
-  "version": "0.1.0-autogenerated.403250971526209557",
+  "version": "0.1.0-autogenerated.11438231719699111413",
   "name": "@polkadot-api/descriptors",
   "files": [
     "dist"
